#1207 remove system warnings when Quad mesh is run in parallel. Note r7789 was by me
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@7791 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/QuadraticMesh.cpp
@@ -242,33 +242,36 @@ void QuadraticMesh<DIM>::RunMesherAndReadMesh(std::string binary,
                                               std::string outputDir,
                                               std::string fileStem)
 {
-
     assert(DIM == 3);
-    std::string args = "-Qo2";
 
-
-    std::string command =  binary + " " + args + " " + outputDir
-                           + "/" + fileStem + ".node" + " > /dev/null";
-
-    int return_value = system(command.c_str());
-
-    if (return_value != 0)
+    if(PetscTools::AmMaster())
     {
-        #define COVERAGE_IGNORE
-        EXCEPTION("Remeshing (by calling " + binary + ") failed.  Do you have it in your path?\n"+
-        "The quadratic mesh relies on functionality from tetgen (http://tetgen.berlios.de/).");
-        #undef COVERAGE_IGNORE
+        std::string args = "-Qo2";
+    
+        std::string command =  binary + " " + args + " " + outputDir
+                           + "/" + fileStem + ".node" + " > /dev/null";
+    
+        int return_value = system(command.c_str());
+    
+        if (return_value != 0)
+        {
+            #define COVERAGE_IGNORE
+            EXCEPTION("Remeshing (by calling " + binary + ") failed.  Do you have it in your path?\n"+
+            "The quadratic mesh relies on functionality from tetgen (http://tetgen.berlios.de/).");
+            #undef COVERAGE_IGNORE
+        }
+    
+        // move the output files to the chaste directory
+        command =   "mv " + outputDir + "/"
+                  + fileStem + ".1.* .";
+    
+        // NOTE: we don't check whether the return value here is zero, because if CHASTE_TESTOUTPUT
+        // is "." (ie if it hasn't been exported), then the mv will fail (source and destination files
+        // are the same), but this isn't a problem.
+        return_value = system(command.c_str());
     }
-
-    // move the output files to the chaste directory
-    command =   "mv " + outputDir + "/"
-              + fileStem + ".1.* .";
-
-    // NOTE: we don't check whether the return value here is zero, because if CHASTE_TESTOUTPUT
-    // is "." (ie if it hasn't been exported), then the mv will fail (source and destination files
-    // are the same), but this isn't a problem.
-    return_value = system(command.c_str());
-
+    PetscTools::Barrier();
+    
     // load
     TrianglesMeshReader<DIM,DIM> mesh_reader(fileStem + ".1", 2, 1, false); // false as tetgen/triangle has been used and therefore boundary elems will be linear;
     ConstructFromMeshReader(mesh_reader);
@@ -276,12 +279,15 @@ void QuadraticMesh<DIM>::RunMesherAndReadMesh(std::string binary,
     // could be a true instead. Currently though there is the intermediate step of having to delete manually the attribute values
     // column after using this flag.
 
-    // delete the temporary files
-    command = "rm -f " + outputDir + "/" + fileStem + ".node";
-    EXPECT0(system, command);
-    EXPECT0(system, "rm -f " + fileStem + ".1.node");
-    EXPECT0(system, "rm -f " + fileStem + ".1.ele");
-    EXPECT0(system, "rm -f " + fileStem + ".1.face");
+    if(PetscTools::AmMaster())
+    {
+        // delete the temporary files
+        std::string command = "rm -f " + outputDir + "/" + fileStem + ".node";
+        EXPECT0(system, command);
+        EXPECT0(system, "rm -f " + fileStem + ".1.node");
+        EXPECT0(system, "rm -f " + fileStem + ".1.ele");
+        EXPECT0(system, "rm -f " + fileStem + ".1.face");
+    }
 }
